move notification items to seperate component
